splitup ephemeral_processing
