+++ src/actinia_core/processing/actinia_processing/ephemeral/raster_colors.py
@@ -28,7 +28,8 @@ TODO: Maybe more tests required, test_raster_colors.py is in place and works
 """
 
 from tempfile import NamedTemporaryFile
-from actinia_core.processing.ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Carmen Tawalika"
