+++ src/actinia_core/rest/persistent_processing.py
@@ -35,7 +35,8 @@ from copy import deepcopy
 from flask import jsonify, make_response
 from flask_restful_swagger_2 import swagger
 
-from actinia_core.rest.ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
