+++ src/actinia_core/rest/process_validation.py
@@ -29,7 +29,8 @@ import pickle
 from copy import deepcopy
 from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response
-from actinia_core.rest.ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.models.response_models import \
