+++ src/actinia_core/rest/raster_legend.py
@@ -27,7 +27,8 @@ Raster map renderer
 
 from flask import jsonify, make_response, Response
 from flask_restful import reqparse
-from .ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from tempfile import NamedTemporaryFile
