+++ src/actinia_core/rest/raster_renderer.py
@@ -29,7 +29,8 @@ from flask_restful_swagger_2 import swagger
 from tempfile import NamedTemporaryFile
 import os
 from flask import jsonify, make_response, Response
-from .ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.rest.base.renderer_base import RendererBaseResource, \
      EphemeralRendererBase
