+++ src/actinia_core/rest/strds_renderer.py
@@ -25,7 +25,8 @@
 Raster map renderer
 """
 from flask import jsonify, make_response, Response
-from .ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.rest.base.renderer_base import RendererBaseResource, \
      EphemeralRendererBase
