+++ src/actinia_core/rest/vector_layer.py
@@ -42,7 +42,8 @@ from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.core.utils import allowed_file
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel, SimpleResponseModel
-from .ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from .persistent_processing import PersistentProcessing
 from actinia_core.rest.base.map_layer_base import MapLayerRegionResourceBase
