+++ src/actinia_core/rest/vector_renderer.py
@@ -30,7 +30,8 @@ import os
 from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response, Response
 from tempfile import NamedTemporaryFile
-from .ephemeral_processing import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.rest.base.renderer_base import RendererBaseResource, \
      EphemeralRendererBase
